"Schedule a call" draft
Added a draft of the "schedule a call" page. As it stands, there is a warning that pops up if the required field is not filled in properly, but this doesn't really satisfy the requirement for user notifications. A modal dialog is probably preferable since according to W3 it's an acceptable way to notify users of errors but I was having an issue so I'll finish it a little later.